update layer classes
